# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Insta-Like API
  description: Photo upload and sharing service
  version: 0.1.0

servers:
  - url: http://localhost:8000
    description: Local development

paths:
  /health:
    get:
      summary: Health check
      description: Returns 200 OK if the service is running.
      operationId: healthCheck
      responses:
        "200":
          description: Service is healthy

  /upload:
    post:
      summary: Upload a photo
      description: Uploads a photo file with optional comma-separated tags. The file is saved to local storage and metadata is persisted to the database.
      operationId: uploadPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - photo
              properties:
                photo:
                  type: string
                  format: binary
                  description: The image file to upload
                tags:
                  type: string
                  description: Comma-separated list of tags (e.g. "sunset, beach, vacation")
                  example: "sunset, beach, vacation"
      responses:
        "201":
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        "400":
          description: Bad request (missing photo field or invalid form)
          content:
            text/plain:
              schema:
                type: string
                example: "missing 'photo' field"
        "500":
          description: Internal server error (file save or database failure)
          content:
            text/plain:
              schema:
                type: string
                example: "failed to save file"

  /photos:
    get:
      summary: List photos
      description: Returns photos from the database with their file content (base64-encoded). Supports cursor-based pagination using the created_at timestamp and optional tag filtering.
      operationId: listPhotos
      parameters:
        - name: tags
          in: query
          required: false
          description: Comma-separated list of tags to filter by. Returns photos matching any of the provided tags.
          schema:
            type: string
          example: "sunset,beach"
        - name: cursor
          in: query
          required: false
          description: Pagination cursor. Pass the next_cursor value from the previous response to fetch the next page. Format is RFC3339Nano timestamp.
          schema:
            type: string
            format: date-time
          example: "2026-05-13T14:30:00.123456789Z"
        - name: limit
          in: query
          required: false
          description: Maximum number of photos to return per page. Defaults to 20.
          schema:
            type: integer
            default: 20
            minimum: 1
          example: 10
      responses:
        "200":
          description: Paginated list of photos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotosResponse"
        "400":
          description: Invalid cursor format
          content:
            text/plain:
              schema:
                type: string
                example: "invalid cursor format, use RFC3339"
        "500":
          description: Internal server error (database or file read failure)
          content:
            text/plain:
              schema:
                type: string
                example: "failed to query photos"

components:
  schemas:
    Photo:
      type: object
      properties:
        Path:
          type: string
          description: File path where the photo is stored
          example: "./uploads/sunset.jpg"
        Tags:
          type: array
          items:
            type: string
          description: List of tags associated with the photo
          example: ["sunset", "beach", "vacation"]
        CreatedAt:
          type: string
          format: date-time
          description: Timestamp when the photo was uploaded
          example: "2026-05-13T14:30:00Z"

    PhotoWithData:
      allOf:
        - $ref: "#/components/schemas/Photo"
        - type: object
          properties:
            Data:
              type: string
              format: byte
              description: Base64-encoded file content. Omitted if the file could not be read.
              example: "/9j/4AAQSkZJRg..."

    PhotosResponse:
      type: object
      properties:
        photos:
          type: array
          items:
            $ref: "#/components/schemas/PhotoWithData"
          description: List of photos for the current page
        next_cursor:
          type: string
          format: date-time
          description: Cursor for the next page. Absent when there are no more results. Pass this value as the cursor query parameter to fetch the next page.
          example: "2026-05-13T14:30:00.123456789Z"